requirements.txt
beartype==0.19.0
contourpy==1.3.1
cycler==0.12.1
deepdiff==8.1.1
fonttools==4.56.0
h5py==3.12.1
jaxtyping==0.2.36
k-wave-python==0.4.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
opencv-python==4.10.0.84
orderly-set==5.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scipy==1.14.1
six==1.17.0
tzdata==2025.1
